requirements.txt
asgiref==3.10.0
attrs==25.4.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-otp==1.6.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
